Convert `Product` price field to decimal
Prices should be decimals for the best accuracy

+++ Domain/Product.cs
@@ -1,9 +1,9 @@
 namespace SimpleInventoryManagementSystem.Domain;
 
-public class Product(string name, int price, int quantity)
+public class Product(string name, decimal price, int quantity)
 {
     public string Name { get; set; } = name;
-    public int Price { get; set; } = price;
+    public decimal Price { get; set; } = price;
     public int Quantity { get; set; } = quantity;
 
     public override string ToString() => $"Product: {Name}, {Price} dollar(s), {Quantity} item(s)";
